+++ uikit/src/main/java/com/sendbird/uikit/fragments/UserTypeListFragment.java
@@ -173,7 +173,7 @@ abstract public class UserTypeListFragment extends BaseGroupChannelFragment impl
     }
 
     private void initChannelUserList(@NonNull GroupChannel channel) {
-        UserTypeListViewModel viewModel = new ViewModelProvider(getActivity(), new ViewModelFactory(channel, customQueryHandler)).get(channel.getUrl(), UserTypeListViewModel.class);
+        UserTypeListViewModel viewModel = new ViewModelProvider(getViewModelStore(), new ViewModelFactory(channel, customQueryHandler)).get(channel.getUrl(), UserTypeListViewModel.class);
         getLifecycle().addObserver(viewModel);
         if (adapter == null) {
             adapter = new UserTypeListAdapter();